Move dialog related stuff to layout/dialogs directory
