+++ components/layout/dialogs/dialog-component.interface.ts
