+++ components/layout/dialogs/dialog-item.class.ts
