+++ components/layout/index.ts
@@ -15,8 +15,8 @@ import mdSidenavDirective from './md-sidenav.directive';
 import mdSwipeAreaDirective from './md-swipe-area.directive';
 import mdToolbarDirective from './md-toolbar.directive';
 import panelCreatorDirective from './panel-creator.directive';
-import {HsDialogContainerComponent} from './dialog-container.component';
-import {HsDialogContainerService} from './dialog-container.service';
+import {HsDialogContainerComponent} from './dialogs/dialog-container.component';
+import {HsDialogContainerService} from './dialogs/dialog-container.service';
 import {HsLayoutService} from './layout.service';
 import {HsPanelHeaderComponent} from './layout-panel-header.component';
 import {downgradeComponent, downgradeInjectable} from '@angular/upgrade/static';
