+++ components/layout/layout.module.ts
@@ -1,8 +1,8 @@
 import {BootstrapComponent} from '../../bootstrap.component';
 import {BrowserModule} from '@angular/platform-browser';
-import {HsDialogContainerComponent} from './dialog-container.component';
-import {HsDialogContainerService} from './dialog-container.service';
-import {HsDialogHostDirective} from './dialog-host.directive';
+import {HsDialogContainerComponent} from './dialogs/dialog-container.component';
+import {HsDialogContainerService} from './dialogs/dialog-container.service';
+import {HsDialogHostDirective} from './dialogs/dialog-host.directive';
 import {HsLayoutServiceProvider} from '../../ajs-upgraded-providers';
 import {NgModule} from '@angular/core';
 /**
