+++ components/save-map/layer-synchronizer.service.ts
@@ -1,7 +1,7 @@
 import Feature from 'ol/Feature';
 import Map from 'ol/Map';
 import {HsCommonEndpointsService} from '../../common/endpoints/endpoints.service';
-import {HsDialogContainerService} from '../layout/dialog-container.service';
+import {HsDialogContainerService} from '../layout/dialogs/dialog-container.service';
 import {HsLaymanService} from './layman.service';
 import {HsSyncErrorDialogComponent} from './sync-error-dialog.component';
 import {HsUtilsService} from '../utils/utils.service';
