+++ components/save-map/sync-error-dialog.component.ts
@@ -1,5 +1,5 @@
 import {Component, Input, ViewRef} from '@angular/core';
-import {HsDialogComponent} from '../layout/dialog-component.interface';
+import {HsDialogComponent} from '../layout/dialogs/dialog-component.interface';
 @Component({
   selector: 'hs-sync-error-dialog',
   template: require('./sync-error-dialog.html'),
